+++ easychart/models/__init__.py
@@ -1,0 +1,4 @@
+from .chart import Chart
+from .grid import Grid
+from .plot import Plot
+from .series import Series
