+++ tests/unit-tests/models.py
@@ -10,7 +10,7 @@ def test_initialization():
 def test_series():
     chart = easychart.new()
     assert isinstance(chart, easychart.Chart)
-    assert isinstance(chart.series, easychart.models.SeriesCollection)
+    assert isinstance(chart.series, easychart.models.Series)
 
 
 def test_chart_show():
